+++ src/lib/Scenes/Map/Bucket.ts
@@ -1,15 +1,11 @@
 import { GlobalMap_viewer } from "__generated__/GlobalMap_viewer.graphql"
 import { filter } from "lodash"
 import moment from "moment"
-import { MapTab } from "./types"
 
-export type BucketKey = "saved" | "fairs" | "galleries" | "museums" | "closing" | "opening"
-export type BucketResults = { [key in BucketKey]: any[] }
-
-export const bucketCityResults = (viewer: GlobalMap_viewer): BucketResults => {
+export const bucketCityResults = (viewer: GlobalMap_viewer) => {
   const saved = filter(viewer.city.shows.edges, e => e.node.is_followed === true)
   const oneWeekFromNow = moment(new Date()).add(1, "week")
-  const fairs = (viewer.city.fairs.edges as unknown) as MapTab[]
+  const fairs = viewer.city.fairs.edges
   const galleries = filter(viewer.city.shows.edges, e => e.node.partner.type === "Gallery")
   const museums = filter(
     viewer.city.shows.edges,
@@ -37,6 +33,9 @@ export const bucketCityResults = (viewer: GlobalMap_viewer): BucketResults => {
     opening,
   }
 }
+
+export type BucketKey = keyof ReturnType<typeof bucketCityResults>
+export type BucketResults = ReturnType<typeof bucketCityResults>
 
 export const emptyBucketResults: BucketResults = {
   saved: [],