+++ src/lib/utils/convertCityToGeoJSON.ts
@@ -1,3 +1,5 @@
+import { GlobalMap_viewer } from "__generated__/GlobalMap_viewer.graphql"
+
 import { MapGeoFeatureCollection } from "lib/Scenes/Map/types"
 
 // Here is a sample GeoJSON document
@@ -15,6 +17,29 @@ import { MapGeoFeatureCollection } from "lib/Scenes/Map/types"
 //         }
 //       },
 // }
+
+export type FairsEdge = GlobalMap_viewer["city"]["fairs"]["edges"]
+export type ShowsEdge = GlobalMap_viewer["city"]["shows"]["edges"]
+
+export const showsToGeoCityShow = (edges: ShowsEdge) =>
+  edges.filter(a => a.node.type === "Show").map(({ node }) => {
+    return {
+      node: {
+        ...node,
+        icon: node.is_followed ? "pin-saved" : "pin",
+      },
+    }
+  })
+
+export const fairToGeoCityFairs = (edges: FairsEdge) =>
+  edges.map(({ node }) => {
+    return {
+      node: {
+        ...node,
+        icon: "pin",
+      },
+    }
+  })
 
 export const convertCityToGeoJSON = data => {
   return {