+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Master
 
+- Shows fairs on the map - orta
 - Update city coordinates (to fix Los Angeles bug) - anandaroop
 - Fixes warning about missing viewer prop - ash
 - Removes touchable feedback on fair cards - ash